% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calcul_buts_moyens}
\alias{calcul_buts_moyens}
\title{Moyenne des buts à domicile par équipe}
\usage{
calcul_buts_moyens(data)
}
\arguments{
\item{data}{Un data.frame de matchs}
}
\value{
Un data.frame avec moyenne des buts
}
\description{
Moyenne des buts à domicile par équipe
}
