% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_football}
\alias{df_football}
\title{Jeu de données de football}
\format{
Un data.frame avec plusieurs variables :
\describe{
\item{Date}{Date du match}
\item{HomeTeam}{Équipe à domicile}
\item{AwayTeam}{Équipe à l'extérieur}
\item{FTHG}{Buts domicile}
\item{FTAG}{Buts extérieur}
\item{FTR}{Résultat du match}
}
}
\source{
Données football
}
\usage{
df_football
}
\description{
Ce jeu de données contient des matchs de football avec résultats et statistiques.
}
\keyword{datasets}
