% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filtrer_resultat}
\alias{filtrer_resultat}
\title{Filtrer les matchs selon le résultat}
\usage{
filtrer_resultat(data, resultat)
}
\arguments{
\item{data}{Un data.frame de matchs}

\item{resultat}{Résultat souhaité ("H", "D", "A")}
}
\value{
Un data.frame filtré
}
\description{
Filtrer les matchs selon le résultat
}
