% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_buts_moyens}
\alias{plot_buts_moyens}
\title{Graphique des buts moyens par équipe}
\usage{
plot_buts_moyens(data)
}
\arguments{
\item{data}{Un data.frame de matchs}
}
\value{
Un graphique ggplot
}
\description{
Graphique des buts moyens par équipe
}
